+++ TradinatorCore/Public/Indicators/EMA.h
@@ -15,13 +15,13 @@ class EMA : public Indicator
 public:
 	EMA() : Indicator() {}
 	EMA(EIndicatorSource source) : Indicator(source) {}
-	EMA(size_t length) : Indicator(length) {}
-	EMA(EIndicatorSource source, size_t length) : Indicator(source, length) {}
-	EMA(size_t length, std::weak_ptr<Security> security) : Indicator(length, security) {}
-	EMA(EIndicatorSource source, size_t length, std::weak_ptr<Security> security) : Indicator(source, length, security) {}
+	EMA(uint64_t length) : Indicator(length) {}
+	EMA(EIndicatorSource source, uint64_t length) : Indicator(source, length) {}
+	EMA(uint64_t length, std::weak_ptr<Security> security) : Indicator(length, security) {}
+	EMA(EIndicatorSource source, uint64_t length, std::weak_ptr<Security> security) : Indicator(source, length, security) {}
 
 	virtual std::vector<std::vector<double>> Calculate() override;
-	void CalculateRaw(const double* input, double* output, size_t window_size, size_t data_size);
+	void CalculateRaw(const double* input, double* output, uint64_t window_size, uint64_t data_size);
 
 
 	virtual std::string GetName() const override { return std::format("Exponential Moving Average ({})", m_length); }
